Changed size_t to long long in marching cubes to be compatible with msvc openmp implementation

+++ mandelbulber2/src/file_mesh.hpp
@@ -87,13 +87,13 @@ public:
 	struct structSaveMeshData
 	{
 		structSaveMeshData() {}
-		structSaveMeshData(std::vector<double> &_vertices, std::vector<size_t> &_polygons,
+		structSaveMeshData(std::vector<double> &_vertices, std::vector<long long> &_polygons,
 			std::vector<sRGB8> &_colorIndices)
 				: vertices(_vertices), polygons(_polygons), colorIndices(_colorIndices)
 		{
 		}
 		std::vector<double> vertices;
-		std::vector<size_t> polygons;
+		std::vector<long long> polygons;
 		std::vector<sRGB8> colorIndices;
 	};
